# Improve SimpleJsonEncoder.escapeString memory usage