write String slices instead of single chars to the Writer. Iterate over the string to be escaped and write a new slice every time a char needs to be escaped.