@fabiojb 2021-05-07
I also created and ran a Java microbenchmark:

Before:
```
Benchmark                                          Mode  Cnt       Score       Error   Units
Benchmak.testEncode                               thrpt   10  231654,956 ± 39245,086   ops/s
Benchmak.testEncode:·gc.alloc.rate                thrpt   10    2982,905 ±   484,834  MB/sec
Benchmak.testEncode:·gc.alloc.rate.norm           thrpt   10   14184,808 ±    98,620    B/op
Benchmak.testEncode:·gc.churn.G1_Eden_Space       thrpt   10    2878,642 ±   506,329  MB/sec
Benchmak.testEncode:·gc.churn.G1_Eden_Space.norm  thrpt   10   13678,036 ±   318,446    B/op
Benchmak.testEncode:·gc.churn.G1_Old_Gen          thrpt   10      49,120 ±     0,249  MB/sec
Benchmak.testEncode:·gc.churn.G1_Old_Gen.norm     thrpt   10     236,215 ±    41,114    B/op
Benchmak.testEncode:·gc.count                     thrpt   10     334,000              counts
Benchmak.testEncode:·gc.time                      thrpt   10     451,000                  ms
```
After:
```
Benchmark                                          Mode  Cnt       Score       Error   Units
Benchmak.testEncode                               thrpt   10  645469,842 ± 12113,391   ops/s
Benchmak.testEncode:·gc.alloc.rate                thrpt   10    2716,148 ±    43,627  MB/sec
Benchmak.testEncode:·gc.alloc.rate.norm           thrpt   10    4633,482 ±    12,607    B/op
Benchmak.testEncode:·gc.churn.G1_Eden_Space       thrpt   10    2356,656 ±   253,340  MB/sec
Benchmak.testEncode:·gc.churn.G1_Eden_Space.norm  thrpt   10    4019,902 ±   416,759    B/op
Benchmak.testEncode:·gc.churn.G1_Old_Gen          thrpt   10     401,142 ±    83,690  MB/sec
Benchmak.testEncode:·gc.churn.G1_Old_Gen.norm     thrpt   10     684,139 ±   139,178    B/op
Benchmak.testEncode:·gc.count                     thrpt   10      71,000              counts
Benchmak.testEncode:·gc.time                      thrpt   10     217,000                  ms

```
Is notable the reduced GC count and time, and increased throughput, almost 3x